Author: Bleep <tomh@inxpress.net> Log message: Merged in u2_10_10_beta06 changes, create branch u2_10_10_beta07
git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/branches/u2_10_10@34 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ include/client.h
@@ -120,7 +120,7 @@ struct Client {
   HandlerType         handler;   /* message index into command table for parsing */
   struct DNSReply*    dns_reply; /* DNS reply used during client registration */
   struct ListingArgs* listing;
-  size_t              max_sendq; /* cached max send queue for client */
+  unsigned int        max_sendq; /* cached max send queue for client */
   unsigned short      lastsq;    /* # 2k blocks when sendqueued called last */
   unsigned short      port;      /* and the remote port# too :-) */
   unsigned char       targets[MAXTARGETS]; /* Hash values of current targets */
